ayaya

+++ ft_memchr.c
@@ -6,14 +6,14 @@
 /*   By: zbentalh <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 18:10:42 by zbentalh          #+#    #+#             */
-/*   Updated: 2022/10/07 18:13:50 by zbentalh         ###   ########.fr       */
+/*   Updated: 2022/10/11 14:46:09 by zbentalh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include <string.h>
 void *ft_memchr(const void *str, int c, size_t n)
 {
 	char *src;
-	int	i;
+	size_t	i;
 	src = (char*)str;
 	i = 0;
 	while(src[i] && i < n)
